requirements.txt
Cython==0.29.32
dill==0.3.6
easydict==1.11
fiftyone==0.19.1
GPUtil==1.4.0
gym==0.26.1
matplotlib==3.5.3
numpy==1.23.3
opencv-python==4.6.0.66
pandas==2.0.0
Pillow==9.2.0
pytorchcv==0.0.67
quadprog==0.1.11
scikit-image==0.20.0
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.2
six==1.16.0
tensorboard==2.15.0
# torch==1.12.0+cu113
torchattacks==3.3.0
torcheval==0.0.6
torchmetrics==0.9.3
torchtnt==0.1.0
# torchvision==0.13.0+cu113
tqdm==4.64.1
wandb==0.15.12